requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
colorama==0.4.4
dj-database-url==0.5.0
Django==3.2.5
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
gunicorn==20.1.0
Pillow==8.3.0
psycopg2-binary==2.9.1
PyJWT==2.1.0
python-dotenv==0.18.0
pytz==2021.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
whitenoise==5.2.0
